requirements.txt
streamlit
transformers
pillow
torch
BlipProcessor
BlipForQuestionAnswering